+++ packages/claude-plugin/scanner/index.js
@@ -1,6 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.ENV_MASK = exports.EnvRedactor = exports.DEFAULT_GUARD_CONFIG = exports.PromptGuard = exports.DEFAULT_CONFIG = exports.SecretScanner = void 0;
+exports.DEFAULT_MCP_CONFIG = exports.McpGuard = exports.ENV_MASK = exports.EnvRedactor = exports.DEFAULT_GUARD_CONFIG = exports.PromptGuard = exports.DEFAULT_CONFIG = exports.SecretScanner = void 0;
 var SecretScanner_1 = require("./SecretScanner");
 Object.defineProperty(exports, "SecretScanner", { enumerable: true, get: function () { return SecretScanner_1.SecretScanner; } });
 Object.defineProperty(exports, "DEFAULT_CONFIG", { enumerable: true, get: function () { return SecretScanner_1.DEFAULT_CONFIG; } });
@@ -10,4 +10,7 @@ Object.defineProperty(exports, "DEFAULT_GUARD_CONFIG", { enumerable: true, get: 
 var EnvRedactor_1 = require("./EnvRedactor");
 Object.defineProperty(exports, "EnvRedactor", { enumerable: true, get: function () { return EnvRedactor_1.EnvRedactor; } });
 Object.defineProperty(exports, "ENV_MASK", { enumerable: true, get: function () { return EnvRedactor_1.ENV_MASK; } });
+var McpGuard_1 = require("./McpGuard");
+Object.defineProperty(exports, "McpGuard", { enumerable: true, get: function () { return McpGuard_1.McpGuard; } });
+Object.defineProperty(exports, "DEFAULT_MCP_CONFIG", { enumerable: true, get: function () { return McpGuard_1.DEFAULT_MCP_CONFIG; } });
 //# sourceMappingURL=index.js.map